@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "description": "Apache Zeppelin web written in ReactJS",
   "engines": {
-    "node": ">=8.9",
+    "node": ">=10.16",
     "npm": ">=6.6"
   },
   "scripts": {
@@ -35,16 +35,20 @@
     "bootstrap": "4.3.1",
     "connected-react-router": "6.4.0",
     "functional-immutable": "1.0.1",
+    "history": "4.9.0",
     "immutable": "4.0.0-rc.12",
     "lodash": "4.17.14",
     "object-assign": "4.1.1",
+    "prop-types": "15.7.2",
     "react": "16.8.6",
     "react-bootstrap": "1.0.0-beta.9",
     "react-dom": "16.8.6",
+    "react-hot-loader": "4.8.4",
     "react-redux": "7.0.3",
     "react-router-dom": "5.0.0",
     "redux": "4.0.1",
     "redux-actions": "2.6.5",
+    "redux-immutable-state-invariant": "2.1.0",
     "redux-logger": "3.0.6",
     "redux-observable": "1.1.0",
     "redux-thunk": "2.3.0",
@@ -78,14 +82,16 @@
     "cssnano": "4.1.10",
     "enzyme": "3.9.0",
     "enzyme-adapter-react-16": "1.13.0",
-    "eslint": "5.16.0",
-    "eslint-plugin-import": "2.17.2",
+    "eslint": "6.0.1",
+    "eslint-config-airbnb": "17.1.1",
+    "eslint-import-resolver-alias": "1.1.2",
+    "eslint-plugin-import": "2.18.1",
+    "eslint-plugin-jsx-a11y": "6.2.3",
     "eslint-plugin-lodash-fp": "2.2.0-a1",
     "eslint-plugin-react": "7.13.0",
-    "eslint-watch": "5.1.2",
+    "eslint-watch": "6.0.0",
     "file-loader": "3.0.1",
     "hard-source-webpack-plugin": "0.13.1",
-    "history": "4.9.0",
     "html-webpack-plugin": "3.2.0",
     "identity-obj-proxy": "3.0.0",
     "jest": "24.8.0",
@@ -97,11 +103,8 @@
     "opn-cli": "4.1.0",
     "postcss-loader": "3.0.0",
     "prompt": "1.0.0",
-    "prop-types": "15.7.2",
     "raf": "3.4.1",
-    "react-hot-loader": "4.8.4",
     "react-test-renderer": "16.8.6",
-    "redux-immutable-state-invariant": "2.1.0",
     "redux-mock-store": "1.5.3",
     "replace": "1.1.0",
     "rimraf": "2.6.3",
@@ -172,34 +175,5 @@
         ]
       }
     }
-  },
-  "eslintConfig": {
-    "root": true,
-    "extends": [
-      "eslint:recommended",
-      "plugin:import/errors",
-      "plugin:import/warnings",
-      "plugin:react/recommended"
-    ],
-    "plugins": [
-      "react"
-    ],
-    "parser": "babel-eslint",
-    "parserOptions": {
-      "ecmaVersion": 6,
-      "sourceType": "module",
-      "ecmaFeatures": {
-        "jsx": true,
-        "experimentalObjectRestSpread": true
-      }
-    },
-    "env": {
-      "es6": true,
-      "browser": true,
-      "node": true,
-      "jquery": true,
-      "jest": true
-    },
-    "globals": {}
   }
 }
