@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 
 const Notebook = () => (
   <div>This is notebook page</div>
