@@ -1,8 +1,8 @@
-import React, { Component } from "react";
-import PropTypes from "prop-types";
-import { ConnectedRouter } from "connected-react-router";
-import { Provider } from "react-redux";
-import App from "./App";
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { ConnectedRouter } from 'connected-react-router';
+import { Provider } from 'react-redux';
+import App from './App';
 
 export default class Root extends Component {
   render() {
@@ -19,5 +19,5 @@ export default class Root extends Component {
 
 Root.propTypes = {
   store: PropTypes.object.isRequired,
-  history: PropTypes.object.isRequired
+  history: PropTypes.object.isRequired,
 };
