@@ -1,3 +1,3 @@
-export const GET_NOTEBOOKS = "GET_NOTEBOOKS";
-export const GET_NOTEBOOKS_SUCCESS = "GET_NOTEBOOKS_SUCCESS";
-export const GET_NOTEBOOKS_FAILURE = "GET_NOTEBOOKS_FAILURE";
+export const GET_NOTEBOOKS = 'GET_NOTEBOOKS';
+export const GET_NOTEBOOKS_SUCCESS = 'GET_NOTEBOOKS_SUCCESS';
+export const GET_NOTEBOOKS_FAILURE = 'GET_NOTEBOOKS_FAILURE';
