@@ -2,13 +2,13 @@ import {
   GET_NOTEBOOKS,
   GET_NOTEBOOKS_SUCCESS,
   // GET_NOTEBOOKS_FAILURE,
-} from "./actionTypes";
+} from './actionTypes';
 
-export const getNotebooks = (/*payload*/) => ({
+export const getNotebooks = (/* payload */) => ({
   type: GET_NOTEBOOKS,
 });
 
-export const getNotebooksSuccess = (payload) => ({
+export const getNotebooksSuccess = payload => ({
   type: GET_NOTEBOOKS_SUCCESS,
   payload,
 });
