@@ -1,4 +1,4 @@
-import Immutable from "immutable";
+import Immutable from 'immutable';
 
 // Make sure to export an immutable object
 export default Immutable.fromJS({
