@@ -1,6 +1,6 @@
-import { combineEpics } from "redux-observable";
+import { combineEpics } from 'redux-observable';
 
-import notebookEpic from "./notebooks/epics.js";
+import notebookEpic from './notebooks/epics';
 
 const rootEpic = combineEpics(
   notebookEpic,
