@@ -1,7 +1,7 @@
-import { combineReducers } from "redux";
-import { connectRouter } from "connected-react-router";
+import { combineReducers } from 'redux';
+import { connectRouter } from 'connected-react-router';
 
-import notebooks from "./notebooks/reducers";
+import notebooks from './notebooks/reducers';
 
 const rootReducer = history => combineReducers({
   router: connectRouter(history),
