@@ -1,13 +1,13 @@
-import {createStore, compose, applyMiddleware} from "redux";
-import reduxImmutableStateInvariant from "redux-immutable-state-invariant";
-import { createBrowserHistory } from "history";
-import { logger } from "redux-logger";
-import { createEpicMiddleware } from "redux-observable";
+import { createStore, compose, applyMiddleware } from 'redux';
+import reduxImmutableStateInvariant from 'redux-immutable-state-invariant';
+import { createBrowserHistory } from 'history';
+import { logger } from 'redux-logger';
+import { createEpicMiddleware } from 'redux-observable';
 
 // "routerMiddleware": the new way of storing route changes with redux middleware since rrV4.
-import { connectRouter, routerMiddleware } from "connected-react-router";
-import rootReducer from "~/dux/rootReducer";
-import rootEpic from "~/dux/rootEpic";
+import { connectRouter, routerMiddleware } from 'connected-react-router';
+import rootReducer from '~/dux/rootReducer';
+import rootEpic from '~/dux/rootEpic';
 
 export const history = createBrowserHistory();
 const connectRouterHistory = connectRouter(history);
@@ -23,7 +23,7 @@ function configureStoreProd(initialState) {
   const store = createStore(
     rootReducer(history), // root reducer with router state
     initialState,
-    compose(applyMiddleware(...middleware))
+    compose(applyMiddleware(...middleware)),
   );
 
   epicMiddleware.run(rootEpic);
@@ -38,22 +38,24 @@ function configureStoreDev(initialState) {
     // Add other middleware on this line...
     logger,
     epicMiddleware,
-    // Redux middleware that spits an error on you when you try to mutate your state either inside a dispatch or between dispatches.
+    // Redux middleware that spits an error on you when you try to
+    // mutate your state either inside a dispatch or between dispatches.
     reduxImmutableStateInvariant(),
     reactRouterMiddleware,
   ];
 
-  const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose; // add support for Redux dev tools
+  // add support for Redux dev tools
+  const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
   const store = createStore(
     rootReducer(history), // root reducer with router state
     initialState,
-    composeEnhancers(applyMiddleware(...middleware))
+    composeEnhancers(applyMiddleware(...middleware)),
   );
   epicMiddleware.run(rootEpic);
   if (module.hot) {
     // Enable Webpack hot module replacement for reducers
-    module.hot.accept("~/dux/rootReducer", () => {
-      const nextRootReducer = require("~/dux/rootReducer").default; // eslint-disable-line global-require
+    module.hot.accept('~/dux/rootReducer', () => {
+      const nextRootReducer = require('~/dux/rootReducer').default; // eslint-disable-line global-require
       store.replaceReducer(connectRouterHistory(nextRootReducer));
     });
   }
@@ -61,6 +63,6 @@ function configureStoreDev(initialState) {
   return store;
 }
 
-const configureStore = process.env.NODE_ENV === "production" ? configureStoreProd : configureStoreDev;
+const configureStore = process.env.NODE_ENV === 'production' ? configureStoreProd : configureStoreDev;
 
 export default configureStore;
