@@ -1,13 +1,15 @@
 // More info on Webpack"s Node API here: https://webpack.js.org/api/node/
 // Allowing console calls below since this is a build file.
 /* eslint-disable no-console */
-import webpack from "webpack";
-import config from "../webpack.config.prod";
-import {chalkError, chalkSuccess, chalkWarning, chalkProcessing} from "./chalkConfig";
+import webpack from 'webpack';
+import config from '../webpack.config.prod';
+import {
+  chalkError, chalkSuccess, chalkWarning, chalkProcessing,
+} from './chalkConfig';
 
-process.env.NODE_ENV = "production"; // this assures React is built in prod mode and that the Babel dev config doesn"t apply.
+process.env.NODE_ENV = 'production'; // this assures React is built in prod mode and that the Babel dev config doesn"t apply.
 
-console.log(chalkProcessing("Generating minified bundle. This will take a moment..."));
+console.log(chalkProcessing('Generating minified bundle. This will take a moment...'));
 
 webpack(config).run((error, stats) => {
   if (error) { // so a fatal error occurred. Stop here.
@@ -18,18 +20,18 @@ webpack(config).run((error, stats) => {
   const jsonStats = stats.toJson();
 
   if (jsonStats.hasErrors) {
-    return jsonStats.errors.map(error => console.log(chalkError(error)));
+    return jsonStats.errors.map(jsonStatsError => console.log(chalkError(jsonStatsError)));
   }
 
   if (jsonStats.hasWarnings) {
-    console.log(chalkWarning("Webpack generated the following warnings: "));
+    console.log(chalkWarning('Webpack generated the following warnings: '));
     jsonStats.warnings.map(warning => console.log(chalkWarning(warning)));
   }
 
   console.log(`Webpack stats: ${stats}`);
 
   // if we got this far, the build succeeded.
-  console.log(chalkSuccess("Your app is compiled in production mode in /dist. It\"s ready to roll!"));
+  console.log(chalkSuccess('Your app is compiled in production mode in /dist. It"s ready to roll!'));
 
   return 0;
 });
