@@ -1,5 +1,6 @@
 // Centralized configuration for chalk, which is used to add color to console.log statements.
-import chalk from "chalk";
+import chalk from 'chalk';
+
 export const chalkError = chalk.red;
 export const chalkSuccess = chalk.green;
 export const chalkWarning = chalk.yellow;
