@@ -3,7 +3,7 @@
 
 import browserSync from 'browser-sync';
 import historyApiFallback from 'connect-history-api-fallback';
-import {chalkProcessing} from './chalkConfig';
+import { chalkProcessing } from './chalkConfig';
 
 /* eslint-disable no-console */
 
@@ -13,15 +13,15 @@ console.log(chalkProcessing('Opening production build...'));
 browserSync({
   port: 4000,
   ui: {
-    port: 4001
+    port: 4001,
   },
   server: {
-    baseDir: 'dist'
+    baseDir: 'dist',
   },
 
   files: [
-    'src/*.html'
+    'src/*.html',
   ],
 
-  middleware: [historyApiFallback()]
+  middleware: [historyApiFallback()],
 });
