@@ -2,14 +2,14 @@
 // which supports hot reloading and synchronized testing.
 
 // Require Browsersync along with webpack and middleware for it
-import browserSync from "browser-sync";
+import browserSync from 'browser-sync';
 // Required for react-router browserHistory
 // see https://github.com/BrowserSync/browser-sync/issues/204#issuecomment-102623643
-import historyApiFallback from "connect-history-api-fallback";
-import webpack from "webpack";
-import webpackDevMiddleware from "webpack-dev-middleware";
-import webpackHotMiddleware from "webpack-hot-middleware";
-import config from "../webpack.config.dev";
+import historyApiFallback from 'connect-history-api-fallback';
+import webpack from 'webpack';
+import webpackDevMiddleware from 'webpack-dev-middleware';
+import webpackHotMiddleware from 'webpack-hot-middleware';
+import config from '../webpack.config.dev';
 
 const bundler = webpack(config);
 
@@ -17,10 +17,10 @@ const bundler = webpack(config);
 browserSync({
   port: 3000,
   ui: {
-    port: 3001
+    port: 3001,
   },
   server: {
-    baseDir: "src",
+    baseDir: 'src',
 
     middleware: [
       historyApiFallback(),
@@ -39,7 +39,7 @@ browserSync({
           hash: false,
           timings: false,
           chunks: false,
-          chunkModules: false
+          chunkModules: false,
         },
 
         // for other settings see
@@ -47,13 +47,13 @@ browserSync({
       }),
 
       // bundler should be the same as above
-      webpackHotMiddleware(bundler)
-    ]
+      webpackHotMiddleware(bundler),
+    ],
   },
 
   // no need to watch "*.js" here, webpack will take care of it for us,
   // including full page reloads if HMR won"t work
   files: [
-    "src/*.html"
-  ]
+    'src/*.html',
+  ],
 });
