@@ -1,5 +1,5 @@
-import {chalkSuccess} from "./chalkConfig";
+import { chalkSuccess } from './chalkConfig';
 
 /* eslint-disable no-console */
 
-console.log(chalkSuccess("Starting app in dev mode..."));
+console.log(chalkSuccess('Starting app in dev mode...'));
