@@ -1,130 +1,131 @@
-import webpack from "webpack";
-import HtmlWebpackPlugin from "html-webpack-plugin";
-import path from "path";
-import HardSourceWebpackPlugin from "hard-source-webpack-plugin";
+import webpack from 'webpack';
+import HtmlWebpackPlugin from 'html-webpack-plugin';
+import path from 'path';
+import HardSourceWebpackPlugin from 'hard-source-webpack-plugin';
 
 export default {
   resolve: {
-    modules: ["node_modules"],
-    extensions: ["*", ".js", ".jsx", ".json"],
+    modules: ['node_modules'],
+    extensions: ['*', '.js', '.jsx', '.json'],
     // To support react-hot-loader
     alias: {
-      "~": path.resolve(__dirname, "./src"),
-      "react-dom": "@hot-loader/react-dom",
-    }
+      '~': path.resolve(__dirname, './src'),
+      'react-dom': '@hot-loader/react-dom',
+    },
   },
-  devtool: "cheap-module-eval-source-map", // more info:https://webpack.js.org/guides/development/#using-source-maps and https://webpack.js.org/configuration/devtool/
+  devtool: 'cheap-module-eval-source-map', // more info:https://webpack.js.org/guides/development/#using-source-maps and https://webpack.js.org/configuration/devtool/
   entry: [
     // must be first entry to properly set public path
-    "./src/webpack-public-path",
-    "react-hot-loader/patch",
-    "webpack-hot-middleware/client?reload=true",
-    path.resolve(__dirname, "src/index.js"), // Defining path seems necessary for this to work consistently on Windows machines.
+    './src/webpack-public-path',
+    'react-hot-loader/patch',
+    'webpack-hot-middleware/client?reload=true',
+    path.resolve(__dirname, 'src/index.jsx'), // Defining path seems necessary for this to work consistently on Windows machines.
   ],
-  target: "web",
-  mode: "development",
+  target: 'web',
+  mode: 'development',
   output: {
-    path: path.resolve(__dirname, "dist"), // Note: Physical files are only output by the production build task `npm run build`.
-    publicPath: "/",
-    filename: "bundle.js"
+    path: path.resolve(__dirname, 'dist'), // Note: Physical files are only output by the production build task `npm run build`.
+    publicPath: '/',
+    filename: 'bundle.js',
   },
   plugins: [
     new HardSourceWebpackPlugin(),
     new webpack.HotModuleReplacementPlugin(),
     new webpack.NoEmitOnErrorsPlugin(),
-    new HtmlWebpackPlugin({     // Create HTML file that includes references to bundled CSS and JS.
-      template: "src/index.ejs",
+    new HtmlWebpackPlugin({ // Create HTML file that includes references to bundled CSS and JS.
+      template: 'src/index.ejs',
       minify: {
         removeComments: true,
-        collapseWhitespace: true
+        collapseWhitespace: true,
       },
-      inject: true
-    })
+      inject: true,
+    }),
   ],
   module: {
     rules: [
       {
         test: /\.jsx?$/,
         exclude: /node_modules/,
-        use: ["babel-loader"]
+        use: ['babel-loader'],
       },
       {
         test: /\.eot(\?v=\d+.\d+.\d+)?$/,
-        use: ["file-loader"]
+        use: ['file-loader'],
       },
       {
         test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
         use: [
           {
-            loader: "url-loader",
+            loader: 'url-loader',
             options: {
               limit: 10000,
-              mimetype: "application/font-woff"
-            }
-          }
-        ]
+              mimetype: 'application/font-woff',
+            },
+          },
+        ],
       },
       {
         test: /\.[ot]tf(\?v=\d+.\d+.\d+)?$/,
         use: [
           {
-            loader: "url-loader",
+            loader: 'url-loader',
             options: {
               limit: 10000,
-              mimetype: "application/octet-stream"
-            }
-          }
-        ]
+              mimetype: 'application/octet-stream',
+            },
+          },
+        ],
       },
       {
         test: /\.svg(\?v=\d+\.\d+\.\d+)?$/,
         use: [
           {
-            loader: "url-loader",
+            loader: 'url-loader',
             options: {
               limit: 10000,
-              mimetype: "image/svg+xml"
-            }
-          }
-        ]
+              mimetype: 'image/svg+xml',
+            },
+          },
+        ],
       },
       {
         test: /\.(jpe?g|png|gif|ico)$/i,
         use: [
           {
-            loader: "file-loader",
+            loader: 'file-loader',
             options: {
-              name: "[name].[ext]"
-            }
-          }
-        ]
+              name: '[name].[ext]',
+            },
+          },
+        ],
       },
       {
         test: /(\.css|\.scss|\.sass)$/,
         use: [
-          "style-loader",
+          'style-loader',
           {
-            loader: "css-loader",
+            loader: 'css-loader',
             options: {
-              sourceMap: true
-            }
+              sourceMap: true,
+            },
           }, {
-            loader: "postcss-loader",
+            loader: 'postcss-loader',
             options: {
               plugins: () => [
-                require("autoprefixer")
+                // eslint-disable-next-line global-require
+                require('autoprefixer'),
               ],
-              sourceMap: true
-            }
+              sourceMap: true,
+            },
           }, {
-            loader: "sass-loader",
+            loader: 'sass-loader',
             options: {
-              includePaths: [path.resolve(__dirname, "src")],
-              sourceMap: true
-            }
-          }
-        ]
-      }
-    ]
-  }
+              includePaths: [path.resolve(__dirname, 'src')],
+              sourceMap: true,
+            },
+          },
+        ],
+      },
+    ],
+  },
 };
